Add routes to update registration

+++ api/src/main/java/com/codeforcommunity/dto/checkout/LineItem.java
@@ -3,8 +3,8 @@ package com.codeforcommunity.dto.checkout;
 public class LineItem {
   private String name;
   private String description;
-  private Long cents;
-  private Long quantity;
+  private Integer cents;
+  private Integer quantity;
   private Integer id;
 
   private LineItem() {}
@@ -12,8 +12,8 @@ public class LineItem {
   public LineItem(String name, String description, Integer amount, Integer quantity, Integer id) {
     this.name = name;
     this.description = description;
-    this.cents = Long.valueOf(amount);
-    this.quantity = Long.valueOf(quantity);
+    this.cents = amount;
+    this.quantity = quantity;
     this.id = id;
   }
 
@@ -25,11 +25,11 @@ public class LineItem {
     return description;
   }
 
-  public Long getCents() {
+  public Integer getCents() {
     return cents;
   }
 
-  public Long getQuantity() {
+  public Integer getQuantity() {
     return quantity;
   }
